xsetwacom: fix compiler warning.
xsetwacom.c: In function ‘list_mod’:
xsetwacom.c:908:2: warning: format ‘%d’ expects type ‘int’, but argument 2
has type ‘long unsigned int’
xsetwacom.c:912:2: warning: format ‘%d’ expects type ‘int’, but argument 2
has type ‘long unsigned int’

Cast ArrayLength to unsigned int, we won't deal with any arrays larger than
UINT_MAX.

+++ tools/xsetwacom.c
@@ -40,7 +40,7 @@
 #define TRACE(...) \
 	if (verbose) fprintf(stderr, "... " __VA_ARGS__)
 
-#define ArrayLength(a) (sizeof(a) / (sizeof((a)[0])))
+#define ArrayLength(a) ((unsigned int)(sizeof(a) / (sizeof((a)[0]))))
 
 static int verbose = False;
